Minor changes to avoid compiler warnings
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15527 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/andorSrc/andorCCD.cpp
@@ -1091,7 +1091,7 @@ void AndorCCD::dataTask(void)
   NDDataType_t dataType;
   at_32 firstImage, lastImage;
   at_32 validFirst, validLast;
-  int dims[2];
+  size_t dims[2];
   int nDims = 2;
   int i;
   epicsTimeStamp startTime;
